pyproject.toml
[project]
name = "mapa-objekty"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.8"
dependencies = [
    "keras>=3.13.2",
    "keras-hub>=0.26.0",
    "matplotlib>=3.10.8",
    "numpy>=2.4.3",
    "opencv-python>=4.13.0.92",
    "pillow>=12.1.1",
    "rasterio>=1.4.4",
    "scikit-learn>=1.8.0",
    "scipy>=1.17.1",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
    "tqdm>=4.67.3",
]
